requirements.txt
# Core dependencies
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# Options data and analysis
py-vollib>=1.0.1
scipy>=1.11.0

# Data visualization (optional)
matplotlib>=3.7.0
plotly>=5.14.0

# Configuration
pyyaml>=6.0

# API clients (optional - uncomment if using)
# alpha-vantage>=2.3.1
# ib-insync>=0.9.86

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
